MEDIA-0001 split rec bb
