+++ CMakeLists.txt
@@ -157,11 +157,15 @@ set(FILES
         bridge/engine/AudioForwarderRewriteAndSendJob.h
         bridge/engine/EncodeJob.cpp
         bridge/engine/EncodeJob.h
+        bridge/engine/AddPacketCacheJob.h
+        bridge/engine/AddPacketCacheJob.cpp
         bridge/engine/Engine.cpp
         bridge/engine/Engine.h
         bridge/engine/EngineAudioStream.h
         bridge/engine/EngineDataStream.h
         bridge/engine/EngineMixer.cpp
+        bridge/engine/EngineMixerBarbell.cpp
+        bridge/engine/EngineMixerRecording.cpp
         bridge/engine/EngineMixer.h
         bridge/engine/EngineStats.h
         bridge/engine/EngineStreamDirector.h
