+++ bridge/engine/EngineMixer.cpp
@@ -2,8 +2,8 @@
 #include "api/DataChannelMessage.h"
 #include "api/DataChannelMessageParser.h"
 #include "bridge/MixerManagerAsync.h"
-#include "bridge/RecordingDescription.h"
 #include "bridge/engine/ActiveMediaList.h"
+#include "bridge/engine/AddPacketCacheJob.h"
 #include "bridge/engine/AudioForwarderReceiveJob.h"
 #include "bridge/engine/AudioForwarderRewriteAndSendJob.h"
 #include "bridge/engine/DiscardReceivedVideoPacketJob.h"
@@ -11,16 +11,9 @@
 #include "bridge/engine/EngineAudioStream.h"
 #include "bridge/engine/EngineBarbell.h"
 #include "bridge/engine/EngineDataStream.h"
-#include "bridge/engine/EngineRecordingStream.h"
 #include "bridge/engine/EngineStreamDirector.h"
 #include "bridge/engine/EngineVideoStream.h"
 #include "bridge/engine/ProcessMissingVideoPacketsJob.h"
-#include "bridge/engine/ProcessUnackedRecordingEventPacketsJob.h"
-#include "bridge/engine/RecordingAudioForwarderSendJob.h"
-#include "bridge/engine/RecordingEventAckReceiveJob.h"
-#include "bridge/engine/RecordingRtpNackReceiveJob.h"
-#include "bridge/engine/RecordingSendEventJob.h"
-#include "bridge/engine/RecordingVideoForwarderSendJob.h"
 #include "bridge/engine/SendRtcpJob.h"
 #include "bridge/engine/SetMaxMediaBitrateJob.h"
 #include "bridge/engine/SsrcWhitelist.h"
@@ -37,11 +30,6 @@
 #include "rtp/RtpHeader.h"
 #include "transport/Transport.h"
 #include "transport/TransportImpl.h"
-#include "transport/recp/RecControlHeader.h"
-#include "transport/recp/RecDominantSpeakerEventBuilder.h"
-#include "transport/recp/RecStartStopEventBuilder.h"
-#include "transport/recp/RecStreamAddedEventBuilder.h"
-#include "transport/recp/RecStreamRemovedEventBuilder.h"
 #include "utils/SimpleJson.h"
 #include "utils/Span.h"
 #include "webrtc/DataChannel.h"
@@ -53,7 +41,6 @@ namespace
 {
 
 const int16_t mixSampleScaleFactor = 4;
-const uint64_t RECUNACK_PROCESS_INTERVAL = 25 * utils::Time::ms;
 
 memory::UniquePacket createGoodBye(uint32_t ssrc, memory::PacketPoolAllocator& allocator)
 {
@@ -153,68 +140,6 @@ private:
     concurrency::SynchronizationContext& _engineSyncContext;
     bridge::MixerManagerAsync& _mixerManager;
     uint32_t _feedbackSsrc;
-};
-
-class FinalizeRecordingOutboundContextJob : public jobmanager::CountedJob
-{
-public:
-    FinalizeRecordingOutboundContextJob(bridge::EngineMixer& mixer,
-        transport::RecordingTransport& transport,
-        size_t recordingStreamIdHash,
-        concurrency::SynchronizationContext& engineSyncContext,
-        uint32_t ssrc,
-        std::atomic_uint32_t& finalizerCounter)
-        : CountedJob(transport.getJobCounter()),
-          _mixer(mixer),
-          _engineSyncContext(engineSyncContext),
-          _recordingStreamIdHash(recordingStreamIdHash),
-          _ssrc(ssrc),
-          _finalizerCounter(finalizerCounter)
-    {
-        ++finalizerCounter;
-    }
-
-    void run() override
-    {
-        if (0 == --_finalizerCounter)
-        {
-            _engineSyncContext.post(
-                [&]() { _mixer.onRecordingOutboundContextFinalized(_recordingStreamIdHash, _ssrc); });
-        }
-    }
-
-private:
-    bridge::EngineMixer& _mixer;
-    concurrency::SynchronizationContext& _engineSyncContext;
-    size_t _recordingStreamIdHash;
-    uint32_t _ssrc;
-    std::atomic_uint32_t& _finalizerCounter;
-};
-
-class AddPacketCacheJob : public jobmanager::CountedJob
-{
-public:
-    AddPacketCacheJob(transport::Transport& transport,
-        bridge::SsrcOutboundContext& outboundContext,
-        bridge::PacketCache* packetCache)
-        : CountedJob(transport.getJobCounter()),
-          _outboundContext(outboundContext),
-          _packetCache(packetCache)
-    {
-    }
-
-    void run() override
-    {
-        if (_outboundContext.packetCache.isSet() && _outboundContext.packetCache.get())
-        {
-            return;
-        }
-        _outboundContext.packetCache.set(_packetCache);
-    }
-
-private:
-    bridge::SsrcOutboundContext& _outboundContext;
-    bridge::PacketCache* _packetCache;
 };
 
 /**
@@ -559,66 +484,6 @@ void EngineMixer::removeStream(const EngineVideoStream* engineVideoStream)
         [this, engineVideoStream]() { _messageListener.asyncVideoStreamRemoved(*this, *engineVideoStream); });
 }
 
-void EngineMixer::addRecordingStream(EngineRecordingStream* engineRecordingStream)
-{
-    if (_engineRecordingStreams.contains(engineRecordingStream->endpointIdHash))
-    {
-        assert(false);
-        return;
-    }
-
-    _engineRecordingStreams.emplace(engineRecordingStream->endpointIdHash, engineRecordingStream);
-}
-
-void EngineMixer::removeRecordingStream(const EngineRecordingStream& engineRecordingStream)
-{
-    logger::debug("Remove recordingStream", _loggableId.c_str());
-    _engineStreamDirector->removeParticipant(engineRecordingStream.endpointIdHash);
-    _engineStreamDirector->removeParticipantPins(engineRecordingStream.endpointIdHash);
-    _engineRecordingStreams.erase(engineRecordingStream.endpointIdHash);
-}
-
-void EngineMixer::updateRecordingStreamModalities(EngineRecordingStream& engineRecordingStream,
-    bool isAudioEnabled,
-    bool isVideoEnabled,
-    bool isScreenSharingEnabled)
-{
-    if (!engineRecordingStream.isReady)
-    {
-        // I think this is very unlikely or even impossible to happen
-        logger::warn("Received a stream update modality but the stream is not ready yet. endpointIdHash %lu",
-            _loggableId.c_str(),
-            engineRecordingStream.endpointIdHash);
-        return;
-    }
-
-    logger::debug("Update recordingStream modalities, stream: %s audio: %s, video: %s",
-        getLoggableId().c_str(),
-        engineRecordingStream.id.c_str(),
-        isAudioEnabled ? "enabled" : "disabled",
-        isVideoEnabled ? "enabled" : "disabled");
-
-    if (engineRecordingStream.isAudioEnabled != isAudioEnabled)
-    {
-        engineRecordingStream.isAudioEnabled = isAudioEnabled;
-        updateRecordingAudioStreams(engineRecordingStream, isAudioEnabled);
-    }
-
-    if (engineRecordingStream.isVideoEnabled != isVideoEnabled)
-    {
-        engineRecordingStream.isVideoEnabled = isVideoEnabled;
-        updateRecordingVideoStreams(engineRecordingStream, SimulcastStream::VideoContentType::VIDEO, isVideoEnabled);
-    }
-
-    if (engineRecordingStream.isScreenSharingEnabled != isScreenSharingEnabled)
-    {
-        engineRecordingStream.isScreenSharingEnabled = isScreenSharingEnabled;
-        updateRecordingVideoStreams(engineRecordingStream,
-            SimulcastStream::VideoContentType::SLIDES,
-            isScreenSharingEnabled);
-    }
-}
-
 void EngineMixer::addDataSteam(EngineDataStream* engineDataStream)
 {
     const auto endpointIdHash = engineDataStream->endpointIdHash;
@@ -657,15 +522,6 @@ void EngineMixer::startTransport(transport::RtcTransport& transport)
     // start on transport allows incoming packets.
     // Postponing this until callbacks has been set and remote ice and dtls has been configured is vital to avoid
     // race condition and sync problems with ice session and srtpclient
-    transport.start();
-    transport.connect();
-}
-
-void EngineMixer::startRecordingTransport(transport::RecordingTransport& transport)
-{
-    logger::debug("Starting recording transport %s", transport.getLoggableId().c_str(), _loggableId.c_str());
-    transport.setDataReceiver(this);
-
     transport.start();
     transport.connect();
 }
@@ -867,166 +723,6 @@ void EngineMixer::addAudioBuffer(const uint32_t ssrc, AudioBuffer* audioBuffer)
 {
     _mixerSsrcAudioBuffers.erase(ssrc);
     _mixerSsrcAudioBuffers.emplace(ssrc, audioBuffer);
-}
-
-void EngineMixer::recordingStart(EngineRecordingStream& stream, const RecordingDescription& desc)
-{
-    auto seq = stream.recordingEventsOutboundContext.sequenceNumber++;
-    auto timestamp = static_cast<uint32_t>(utils::Time::getAbsoluteTime() / 1000000ULL);
-
-    for (const auto& transportEntry : stream.transports)
-    {
-        auto unackedPacketsTrackerItr = stream.recEventUnackedPacketsTracker.find(transportEntry.first);
-        if (unackedPacketsTrackerItr == stream.recEventUnackedPacketsTracker.end())
-        {
-            logger::error("RecEvent packet tracker not found. Unable to send start recording event to transport %s",
-                _loggableId.c_str(),
-                transportEntry.second.getLoggableId().c_str());
-            continue;
-        }
-
-        auto packet = recp::RecStartStopEventBuilder(_sendAllocator)
-                          .setAudioEnabled(desc.isAudioEnabled)
-                          .setVideoEnabled(desc.isVideoEnabled)
-                          .setSequenceNumber(seq)
-                          .setTimestamp(timestamp)
-                          .setRecordingId(desc.recordingId)
-                          .setUserId(desc.ownerId)
-                          .build();
-        if (!packet)
-        {
-            // This need to be improved. If we can't allocate this event, the recording
-            // must fail. We have to find a way to report this glitch
-            logger::error("No space available to allocate rec start event", _loggableId.c_str());
-            continue;
-        }
-
-        logger::info("Sent recording start event for recordingId %s, streamIdHash %zu, transport %zu",
-            _loggableId.c_str(),
-            desc.recordingId.c_str(),
-            stream.endpointIdHash,
-            transportEntry.first);
-
-        transportEntry.second.getJobQueue().addJob<RecordingSendEventJob>(std::move(packet),
-            transportEntry.second,
-            stream.recordingEventsOutboundContext.packetCache,
-            unackedPacketsTrackerItr->second);
-    }
-}
-
-void EngineMixer::stopRecording(EngineRecordingStream& stream, const RecordingDescription& desc)
-{
-    const auto sequenceNumber = stream.recordingEventsOutboundContext.sequenceNumber++;
-    const auto timestamp = static_cast<uint32_t>(utils::Time::getAbsoluteTime() / 1000000ULL);
-
-    for (const auto& transportEntry : stream.transports)
-    {
-        auto unackedPacketsTrackerItr = stream.recEventUnackedPacketsTracker.find(transportEntry.first);
-        if (unackedPacketsTrackerItr == stream.recEventUnackedPacketsTracker.end())
-        {
-            logger::error("RecEvent packet tracker not found. Unable to send stop recording event to %s",
-                _loggableId.c_str(),
-                transportEntry.second.getLoggableId().c_str());
-            continue;
-        }
-
-        auto packet = recp::RecStartStopEventBuilder(_sendAllocator)
-                          .setAudioEnabled(false)
-                          .setVideoEnabled(false)
-                          .setSequenceNumber(sequenceNumber)
-                          .setTimestamp(timestamp)
-                          .setRecordingId(desc.recordingId)
-                          .setUserId(desc.ownerId)
-                          .build();
-
-        if (!packet)
-        {
-            // This need to be improved. If we can't allocate this event, the recording
-            // must fail as we will not know when it finish. We have to find a way to report this glitch
-            logger::error("No space available to allocate rec stop event", _loggableId.c_str());
-            continue;
-        }
-
-        logger::info("Sent recording stop event for recordingId %s, streamIdHash %zu, transport %zu",
-            _loggableId.c_str(),
-            desc.recordingId.c_str(),
-            stream.endpointIdHash,
-            transportEntry.first);
-
-        transportEntry.second.getJobQueue().addJob<RecordingSendEventJob>(std::move(packet),
-            transportEntry.second,
-            stream.recordingEventsOutboundContext.packetCache,
-            unackedPacketsTrackerItr->second);
-    }
-}
-
-void EngineMixer::addRecordingRtpPacketCache(const uint32_t ssrc, const size_t endpointIdHash, PacketCache* packetCache)
-{
-    assert(endpointIdHash);
-
-    auto recordingStream = _engineRecordingStreams.getItem(endpointIdHash);
-    if (!recordingStream)
-    {
-        return;
-    }
-
-    auto* outboundContext = recordingStream->ssrcOutboundContexts.getItem(ssrc);
-    if (outboundContext)
-    {
-        if (outboundContext->markedForDeletion)
-        {
-            logger::warn("Outbound context for recording video packet cache is marked for deletion. ssrc %u",
-                _loggableId.c_str(),
-                ssrc);
-            return;
-            return;
-        }
-
-        auto* transport = recordingStream->transports.getItem(endpointIdHash);
-        if (!transport)
-        {
-            logger::warn("cannot find transport and outbound context for recording video packet cache. ssrc %u",
-                _loggableId.c_str(),
-                ssrc);
-            return;
-        }
-
-        transport->getJobQueue().addJob<AddPacketCacheJob>(*transport, *outboundContext, packetCache);
-    }
-}
-
-void EngineMixer::addTransportToRecordingStream(const size_t streamIdHash,
-    transport::RecordingTransport& transport,
-    UnackedPacketsTracker& recUnackedPacketsTracker)
-{
-    auto* recordingStream = _engineRecordingStreams.getItem(streamIdHash);
-    if (!recordingStream)
-    {
-        return;
-    }
-
-    recordingStream->transports.emplace(transport.getEndpointIdHash(), transport);
-    recordingStream->recEventUnackedPacketsTracker.emplace(transport.getEndpointIdHash(), recUnackedPacketsTracker);
-
-    if (!recordingStream->isReady)
-    {
-        recordingStream->isReady = true;
-        startRecordingAllCurrentStreams(*recordingStream);
-    }
-}
-
-void EngineMixer::removeTransportFromRecordingStream(const size_t streamIdHash, const size_t endpointIdHash)
-{
-    auto* recordingStream = _engineRecordingStreams.getItem(streamIdHash);
-    if (!recordingStream)
-    {
-        return;
-    }
-
-    recordingStream->transports.erase(endpointIdHash);
-    recordingStream->recEventUnackedPacketsTracker.erase(endpointIdHash);
-
-    _messageListener.asyncRemoveRecordingTransport(*this, recordingStream->id.c_str(), endpointIdHash);
 }
 
 void EngineMixer::clear()
@@ -1311,93 +1007,6 @@ void EngineMixer::checkInboundPacketCounters(const uint64_t timestamp)
     }
 }
 
-void EngineMixer::checkRecordingPacketCounters(const uint64_t timestamp)
-{
-    // We will give 15 seconds before delete outbound context since it was marked for deletion
-    // because we can still receive some NACK or need to resend some recording events if smb hasn't
-    // receive the ACKs. The recording events are especially important to recording to be converted properly
-    static constexpr uint64_t decommissionedTimeout = 15 * utils::Time::sec;
-    for (auto& recordingStreamEntry : _engineRecordingStreams)
-    {
-        const auto recordingStreamIdHash = recordingStreamEntry.first;
-        for (auto& outboundContextEntry : recordingStreamEntry.second->ssrcOutboundContexts)
-        {
-            auto& outboundContext = outboundContextEntry.second;
-
-            const bool isDecommissioned =
-                outboundContext.recordingOutboundDecommissioned && !outboundContext.markedForDeletion;
-
-            if (isDecommissioned && utils::Time::diffGT(outboundContext.lastSendTime, timestamp, decommissionedTimeout))
-            {
-                logger::info("Marking for deletion decommissioned recording outbound context ssrc %u, rec "
-                             "endpointIdHash %zu",
-                    _loggableId.c_str(),
-                    outboundContextEntry.first,
-                    recordingStreamIdHash);
-
-                const auto ssrc = outboundContextEntry.first;
-                outboundContext.markedForDeletion = true;
-                auto& outboundContextsFinalizerCounter = recordingStreamEntry.second->outboundContextsFinalizerCounter;
-                auto counterIt = outboundContextsFinalizerCounter.find(ssrc);
-                if (counterIt == outboundContextsFinalizerCounter.end())
-                {
-                    auto pair = outboundContextsFinalizerCounter.emplace(ssrc, 0);
-                    if (!pair.second)
-                    {
-                        logger::error("No space to emplace recording counter for endpoint %zu, ssrc %u",
-                            _loggableId.c_str(),
-                            recordingStreamIdHash,
-                            ssrc);
-                        outboundContext.markedForDeletion = false;
-                        // This will reschedule new execution in 5 seconds
-                        outboundContext.lastSendTime = timestamp - decommissionedTimeout + 5 * utils::Time::sec;
-                        return;
-                    }
-
-                    counterIt = pair.first;
-                }
-
-                std::atomic_uint32_t& finalizeCounter = counterIt->second;
-                for (auto& transportEntry : recordingStreamEntry.second->transports)
-                {
-                    transportEntry.second.getJobQueue().addJob<FinalizeRecordingOutboundContextJob>(*this,
-                        transportEntry.second,
-                        recordingStreamIdHash,
-                        _engineSyncContext,
-                        ssrc,
-                        finalizeCounter);
-                }
-            }
-        }
-    }
-}
-
-void EngineMixer::checkBarbellPacketCounters(const uint64_t timestamp)
-{
-    for (auto& barbellIt : _engineBarbells)
-    {
-        auto barbell = barbellIt.second;
-        const auto packetCount = barbell->transport.getInboundPacketCount();
-        if (barbell->inboundPackets.timestamp == 0)
-        {
-            barbell->inboundPackets.count = packetCount;
-            barbell->inboundPackets.timestamp = timestamp;
-        }
-        else if (utils::Time::diffGE(barbell->inboundPackets.timestamp, timestamp, utils::Time::sec * 15))
-        {
-            if (barbell->inboundPackets.count == packetCount)
-            {
-                removeBarbell(barbell->idHash);
-            }
-            else
-            {
-                barbell->inboundPackets.count = packetCount;
-                barbell->inboundPackets.timestamp = timestamp;
-            }
-        }
-    }
-}
-
 EngineStats::MixerStats EngineMixer::gatherStats(const uint64_t iterationStartTime)
 {
     EngineStats::MixerStats stats;
@@ -1538,23 +1147,6 @@ void EngineMixer::onVideoRtpPacketReceived(SsrcInboundContext& ssrcContext,
         _localVideoSsrc,
         extendedSequenceNumber,
         timestamp);
-}
-
-utils::Optional<uint32_t> EngineMixer::findBarbellMainSsrc(size_t barbellIdHash, uint32_t feedbackSsrc)
-{
-    auto barbell = _engineBarbells.getItem(barbellIdHash);
-    if (!barbell)
-    {
-        return utils::Optional<uint32_t>();
-    }
-
-    auto* videoStream = barbell->videoSsrcMap.getItem(feedbackSsrc);
-    if (!videoStream)
-    {
-        assert(false);
-        return utils::Optional<uint32_t>();
-    }
-    return videoStream->stream.getMainSsrcFor(feedbackSsrc);
 }
 
 void EngineMixer::onVideoRtpRtxPacketReceived(SsrcInboundContext& rtxSsrcContext,
@@ -1800,77 +1392,6 @@ void EngineMixer::onSctpMessage(transport::RtcTransport* sender,
     _messageListener.asyncSctpReceived(*this, packet, sender->getEndpointIdHash());
 }
 
-void EngineMixer::onRecControlReceived(transport::RecordingTransport* sender,
-    memory::UniquePacket packet,
-    uint64_t timestamp)
-{
-    auto* recordingStream = _engineRecordingStreams.getItem(sender->getStreamIdHash());
-    if (!recordingStream)
-    {
-        return;
-    }
-
-    auto recControlHeader = recp::RecControlHeader::fromPacket(*packet);
-    if (recControlHeader->isEventAck())
-    {
-        auto* unackedPacketsTracker =
-            recordingStream->recEventUnackedPacketsTracker.getItem(sender->getEndpointIdHash());
-        if (!unackedPacketsTracker)
-        {
-            return;
-        }
-
-        sender->getJobQueue().addJob<RecordingEventAckReceiveJob>(std::move(packet), sender, *unackedPacketsTracker);
-    }
-    else if (recControlHeader->isRtpNack())
-    {
-        auto ssrc = recControlHeader->getSsrc();
-        auto ssrcContext = recordingStream->ssrcOutboundContexts.getItem(ssrc);
-        if (!ssrcContext || ssrcContext->markedForDeletion)
-        {
-            return;
-        }
-
-        sender->getJobQueue().addJob<RecordingRtpNackReceiveJob>(std::move(packet),
-            _sendAllocator,
-            sender,
-            *ssrcContext);
-    }
-}
-
-bool EngineMixer::setPacketSourceEndpointIdHash(memory::Packet& packet,
-    size_t barbellIdHash,
-    uint32_t ssrc,
-    bool isAudio)
-{
-    auto* barbell = _engineBarbells.getItem(barbellIdHash);
-    if (!barbell)
-    {
-        return false;
-    }
-
-    if (isAudio)
-    {
-        const auto* audioStream = barbell->audioSsrcMap.getItem(ssrc);
-        if (audioStream && audioStream->endpointIdHash.isSet())
-        {
-            packet.endpointIdHash = audioStream->endpointIdHash.get();
-            return true;
-        }
-    }
-    else
-    {
-        const auto* videoStream = barbell->videoSsrcMap.getItem(ssrc);
-        if (videoStream && videoStream->endpointIdHash.isSet())
-        {
-            packet.endpointIdHash = videoStream->endpointIdHash.get();
-            return true;
-        }
-    }
-
-    return false;
-}
-
 /**
  * We store source endpointIdHash in the Packet itself, because, after arriving here, the packet goes through
  * decryption and is ready to be forwarded. At that time we want to know how this source endpoint is mapped to
@@ -2083,62 +1604,6 @@ void EngineMixer::onOutboundContextFinalized(size_t ownerEndpointHash,
     }
 }
 
-void EngineMixer::onRecordingOutboundContextFinalized(size_t recordingStreamIdHash, uint32_t ssrc)
-{
-    auto* recordingStream = _engineRecordingStreams.getItem(recordingStreamIdHash);
-    if (!recordingStream)
-    {
-        return;
-    }
-
-    auto& outboundContexts = recordingStream->ssrcOutboundContexts;
-    const auto outboundContextIt = outboundContexts.find(ssrc);
-    if (outboundContextIt == outboundContexts.end())
-    {
-        return;
-    }
-
-    // It can happen that outbound context can be re-activated after marked for deletion. Then we will check if it
-    // is still marked for deletion.
-    if (!outboundContextIt->second.markedForDeletion)
-    {
-        logger::info("Recording outbound context was re-activated for endpointId %zu, ssrc %u",
-            _loggableId.c_str(),
-            recordingStreamIdHash,
-            ssrc);
-
-        recordingStream->outboundContextsFinalizerCounter.erase(ssrc);
-        return;
-    }
-
-    auto& outboundContextsFinalizerCounter = recordingStream->outboundContextsFinalizerCounter;
-    auto finalizeCounterIt = outboundContextsFinalizerCounter.find(ssrc);
-    assert(finalizeCounterIt != outboundContextsFinalizerCounter.end());
-    if (finalizeCounterIt != outboundContextsFinalizerCounter.end())
-    {
-        // It's allowed to re-active outbound ssrc, even though this should never happen.
-        // but for safety we will verify if a reactivation has provoked new finalized jobs
-        if (finalizeCounterIt->second != 0)
-        {
-            logger::warn("Finalize counter is not zero yet for endpoint %zu, ssrc %u",
-                _loggableId.c_str(),
-                recordingStreamIdHash,
-                ssrc);
-            return;
-        }
-
-        outboundContextsFinalizerCounter.erase(ssrc);
-    }
-
-    logger::info("Removing recording outbound context for endpoint %zu, ssrc %u",
-        _loggableId.c_str(),
-        recordingStreamIdHash,
-        ssrc);
-
-    _messageListener.asyncFreeRecordingRtpPacketCache(*this, outboundContextIt->second.ssrc, recordingStreamIdHash);
-    outboundContexts.erase(outboundContextIt->first);
-}
-
 SsrcInboundContext* EngineMixer::emplaceInboundSsrcContext(const uint32_t ssrc,
     transport::RtcTransport* sender,
     const uint32_t payloadType,
@@ -2297,113 +1762,6 @@ SsrcInboundContext* EngineMixer::emplaceInboundSsrcContext(const uint32_t ssrc,
     return nullptr;
 }
 
-SsrcInboundContext* EngineMixer::emplaceBarbellInboundSsrcContext(const uint32_t ssrc,
-    transport::RtcTransport* sender,
-    const uint32_t payloadType,
-    const uint64_t timestamp)
-{
-    const auto endpointIdHash = sender->getEndpointIdHash();
-    auto* barbell = _engineBarbells.getItem(endpointIdHash);
-    if (!barbell)
-    {
-        return nullptr;
-    }
-
-    auto videoStream = barbell->videoSsrcMap.getItem(ssrc);
-    if (videoStream)
-    {
-        const RtpMap& videoRtpMap =
-            barbell->videoRtpMap.payloadType == payloadType ? barbell->videoRtpMap : barbell->videoFeedbackRtpMap;
-        auto simulcastLevel = videoStream->stream.getLevelOf(ssrc);
-        if (barbell->videoRtpMap.payloadType == payloadType && !simulcastLevel.isSet())
-        {
-            logger::error("ssrc %u is not in simulcast group of barbell video stream %zu",
-                _loggableId.c_str(),
-                ssrc,
-                barbell->idHash);
-            return nullptr;
-        }
-
-        auto emplaceResult = _allSsrcInboundContexts.emplace(ssrc,
-            ssrc,
-            videoRtpMap,
-            sender,
-            timestamp,
-            simulcastLevel.get(),
-            videoStream->stream.levels[0].ssrc);
-        if (!emplaceResult.second && emplaceResult.first == _allSsrcInboundContexts.end())
-        {
-            logger::error("Failed to create barbell inbound video context for ssrc %u", _loggableId.c_str(), ssrc);
-            return nullptr;
-        }
-
-        _ssrcInboundContexts.emplace(ssrc, &emplaceResult.first->second);
-
-        logger::info("Created new barbell inbound video context for stream ssrc %u, endpointIdHash %zu, %s",
-            _loggableId.c_str(),
-            ssrc,
-            endpointIdHash,
-            sender->getLoggableId().c_str());
-        return &emplaceResult.first->second;
-    }
-
-    if (barbell->audioSsrcMap.contains(ssrc))
-    {
-        auto emplaceResult = _allSsrcInboundContexts.emplace(ssrc, ssrc, barbell->audioRtpMap, sender, timestamp);
-
-        if (!emplaceResult.second && emplaceResult.first == _allSsrcInboundContexts.end())
-        {
-            logger::error("Failed to create barbell inbound audio context for ssrc %u", _loggableId.c_str(), ssrc);
-            return nullptr;
-        }
-        _ssrcInboundContexts.emplace(ssrc, &emplaceResult.first->second);
-
-        logger::info("Created new barbell inbound audio context for stream ssrc %u, endpointIdHash %zu, %s",
-            _loggableId.c_str(),
-            ssrc,
-            endpointIdHash,
-            sender->getLoggableId().c_str());
-        return &emplaceResult.first->second;
-    }
-
-    return nullptr;
-}
-
-void EngineMixer::processBarbellSctp(const uint64_t timestamp)
-{
-    for (IncomingPacketInfo packetInfo; _incomingBarbellSctp.pop(packetInfo);)
-    {
-        auto header = reinterpret_cast<webrtc::SctpStreamMessageHeader*>(packetInfo.packet()->get());
-
-        if (header->payloadProtocol == webrtc::DataChannelPpid::WEBRTC_STRING)
-        {
-            auto message = reinterpret_cast<const char*>(header->data());
-            const auto messageLength = header->getMessageLength(packetInfo.packet()->getLength());
-            if (messageLength == 0)
-            {
-                return;
-            }
-
-            auto messageJson = utils::SimpleJson::create(message, messageLength);
-
-            if (api::DataChannelMessageParser::isUserMediaMap(messageJson))
-            {
-                onBarbellUserMediaMap(packetInfo.transport()->getEndpointIdHash(), message);
-            }
-            else if (api::DataChannelMessageParser::isMinUplinkBitrate(messageJson))
-            {
-                onBarbellMinUplinkEstimate(packetInfo.transport()->getEndpointIdHash(), message);
-            }
-        }
-        else if (header->payloadProtocol == webrtc::DataChannelPpid::WEBRTC_ESTABLISH)
-        {
-            onBarbellDataChannelEstablish(packetInfo.transport()->getEndpointIdHash(),
-                *header,
-                packetInfo.packet()->getLength());
-        }
-    }
-}
-
 namespace
 {
 template <class V, class T>
@@ -2535,92 +1893,6 @@ void EngineMixer::forwardAudioRtpPacket(IncomingPacketInfo& packetInfo, uint64_t
     }
 }
 
-void EngineMixer::forwardAudioRtpPacketRecording(IncomingPacketInfo& packetInfo, uint64_t timestamp)
-{
-    if (EngineBarbell::isFromBarbell(packetInfo.transport()->getTag()) || !packetInfo.inboundContext())
-    {
-        return;
-    }
-
-    for (auto& recordingStreams : _engineRecordingStreams)
-    {
-        auto* recordingStream = recordingStreams.second;
-        if (!(recordingStream && recordingStream->isAudioEnabled))
-        {
-            continue;
-        }
-
-        const auto ssrc = packetInfo.inboundContext()->ssrc;
-        auto* ssrcOutboundContext = recordingStream->ssrcOutboundContexts.getItem(ssrc);
-        if (!ssrcOutboundContext || ssrcOutboundContext->recordingOutboundDecommissioned)
-        {
-            continue;
-        }
-
-        for (const auto& transportEntry : recordingStream->transports)
-        {
-            ssrcOutboundContext->onRtpSent(timestamp);
-            auto packet = memory::makeUniquePacket(_sendAllocator, *packetInfo.packet());
-            if (packet)
-            {
-                transportEntry.second.getJobQueue().addJob<RecordingAudioForwarderSendJob>(*ssrcOutboundContext,
-                    std::move(packet),
-                    transportEntry.second,
-                    packetInfo.extendedSequenceNumber(),
-                    _messageListener,
-                    transportEntry.first,
-                    *this);
-            }
-            else
-            {
-                logger::warn("send allocator depleted RecFwdSend", _loggableId.c_str());
-            }
-        }
-    }
-}
-
-void EngineMixer::forwardAudioRtpPacketOverBarbell(IncomingPacketInfo& packetInfo, uint64_t timestamp)
-{
-    if (EngineBarbell::isFromBarbell(packetInfo.transport()->getTag()))
-    {
-        return;
-    }
-
-    for (auto& it : _engineBarbells)
-    {
-        auto& barbell = *it.second;
-        const auto& audioSsrcRewriteMap = _activeMediaList->getAudioSsrcRewriteMap();
-        const auto rewriteMapItr = audioSsrcRewriteMap.find(packetInfo.packet()->endpointIdHash);
-        if (rewriteMapItr == audioSsrcRewriteMap.end())
-        {
-            continue;
-        }
-        uint32_t ssrc = rewriteMapItr->second;
-
-        auto* ssrcOutboundContext =
-            obtainOutboundSsrcContext(barbell.idHash, barbell.ssrcOutboundContexts, ssrc, barbell.audioRtpMap);
-
-        if (!ssrcOutboundContext)
-        {
-            continue;
-        }
-
-        auto packet = memory::makeUniquePacket(_sendAllocator, *packetInfo.packet());
-        if (packet)
-        {
-            barbell.transport.getJobQueue().addJob<AudioForwarderRewriteAndSendJob>(*ssrcOutboundContext,
-                *(packetInfo.inboundContext()),
-                std::move(packet),
-                packetInfo.extendedSequenceNumber(),
-                barbell.transport);
-        }
-        else
-        {
-            logger::warn("send allocator depleted. forwardAudioRtpPacketOverBarbell", _loggableId.c_str());
-        }
-    }
-}
-
 /**
  * Append RTP audio to pre-buffer for this ssrc in _mixerSsrcAudioBuffers
  */
@@ -2733,67 +2005,6 @@ void EngineMixer::processIncomingRtpPackets(const uint64_t timestamp)
     else
     {
         _lastReceiveTime = timestamp;
-    }
-}
-
-void EngineMixer::forwardVideoRtpPacketOverBarbell(IncomingPacketInfo& packetInfo, const uint64_t timestamp)
-{
-    if (EngineBarbell::isFromBarbell(packetInfo.transport()->getTag()) || !packetInfo.inboundContext())
-    {
-        return;
-    }
-
-    const auto senderEndpointIdHash = packetInfo.packet()->endpointIdHash;
-    for (auto& it : _engineBarbells)
-    {
-        auto& barbell = *it.second;
-
-        uint32_t targetSsrc = 0;
-        const auto simulcastLevel = packetInfo.inboundContext()->simulcastLevel;
-        const auto& screenShareSsrcMapping = _activeMediaList->getVideoScreenShareSsrcMapping();
-
-        if (screenShareSsrcMapping.isSet() && screenShareSsrcMapping.get().first == senderEndpointIdHash &&
-            screenShareSsrcMapping.get().second.ssrc == packetInfo.inboundContext()->ssrc)
-        {
-            targetSsrc = screenShareSsrcMapping.get().second.rewriteSsrc;
-        }
-        else
-        {
-            const auto& videoSsrcRewriteMap = _activeMediaList->getVideoSsrcRewriteMap();
-            const auto* rewriteMapping = videoSsrcRewriteMap.getItem(senderEndpointIdHash);
-            if (!rewriteMapping)
-            {
-                continue;
-            }
-
-            targetSsrc = (*rewriteMapping)[simulcastLevel].main;
-        }
-
-        auto* ssrcOutboundContext =
-            obtainOutboundSsrcContext(barbell.idHash, barbell.ssrcOutboundContexts, targetSsrc, barbell.videoRtpMap);
-
-        if (!ssrcOutboundContext)
-        {
-            continue;
-        }
-
-        ssrcOutboundContext->onRtpSent(timestamp); // marks that we have active jobs on this ssrc context
-        auto packet = memory::makeUniquePacket(_sendAllocator, *packetInfo.packet());
-        if (packet)
-        {
-            barbell.transport.getJobQueue().addJob<VideoForwarderRewriteAndSendJob>(*ssrcOutboundContext,
-                *(packetInfo.inboundContext()),
-                std::move(packet),
-                barbell.transport,
-                packetInfo.extendedSequenceNumber(),
-                _messageListener,
-                barbell.idHash,
-                *this);
-        }
-        else
-        {
-            logger::warn("send allocator depleted fwdVideoOverBarbell", _loggableId.c_str());
-        }
     }
 }
 
@@ -2930,57 +2141,6 @@ void EngineMixer::forwardVideoRtpPacket(IncomingPacketInfo& packetInfo, const ui
     }
 }
 
-void EngineMixer::forwardVideoRtpPacketRecording(IncomingPacketInfo& packetInfo, const uint64_t timestamp)
-{
-    if (EngineBarbell::isFromBarbell(packetInfo.transport()->getTag()) || !packetInfo.inboundContext())
-    {
-        return;
-    }
-
-    for (auto& recordingStreams : _engineRecordingStreams)
-    {
-        auto* recordingStream = recordingStreams.second;
-        if (!(recordingStream && (recordingStream->isVideoEnabled || recordingStream->isScreenSharingEnabled)))
-        {
-            continue;
-        }
-
-        if (!_engineStreamDirector->shouldRecordSsrc(recordingStream->endpointIdHash,
-                packetInfo.inboundContext()->ssrc))
-        {
-            continue;
-        }
-
-        auto* ssrcOutboundContext =
-            recordingStream->ssrcOutboundContexts.getItem(packetInfo.inboundContext()->defaultLevelSsrc);
-        if (!ssrcOutboundContext || ssrcOutboundContext->recordingOutboundDecommissioned)
-        {
-            continue;
-        }
-
-        for (const auto& transportEntry : recordingStream->transports)
-        {
-            ssrcOutboundContext->onRtpSent(timestamp); // active jobs on this ssrc context
-            auto packet = memory::makeUniquePacket(_sendAllocator, *packetInfo.packet());
-            if (packet)
-            {
-                transportEntry.second.getJobQueue().addJob<RecordingVideoForwarderSendJob>(*ssrcOutboundContext,
-                    *(packetInfo.inboundContext()),
-                    std::move(packet),
-                    transportEntry.second,
-                    packetInfo.extendedSequenceNumber(),
-                    _messageListener,
-                    transportEntry.first,
-                    *this);
-            }
-            else
-            {
-                logger::warn("send allocator depleted FwdRewrite", _loggableId.c_str());
-            }
-        }
-    }
-}
-
 void EngineMixer::processIncomingRtcpPackets(const uint64_t timestamp)
 {
     for (IncomingPacketInfo packetInfo; _incomingRtcp.pop(packetInfo);)
@@ -3055,57 +2215,6 @@ void EngineMixer::processIncomingPayloadSpecificRtcpPacket(const size_t rtcpSend
             rtcpSenderEndpointIdHash,
             inboundContext->ssrc);
         inboundContext->pliScheduler.triggerPli();
-    }
-}
-
-void EngineMixer::processIncomingBarbellFbRtcpPacket(EngineBarbell& barbell,
-    const rtp::RtcpFeedback& rtcpFeedback,
-    const uint64_t timestamp)
-{
-    uint32_t rtxSsrc = 0;
-    const auto mediaSsrc = rtcpFeedback.mediaSsrc.get();
-    _activeMediaList->getFeedbackSsrc(mediaSsrc, rtxSsrc);
-
-    auto& bbTransport = barbell.transport;
-    auto* mediaSsrcOutboundContext =
-        obtainOutboundSsrcContext(barbell.idHash, barbell.ssrcOutboundContexts, mediaSsrc, barbell.videoRtpMap);
-    if (!mediaSsrcOutboundContext)
-    {
-        return;
-    }
-
-    auto* rtxSsrcOutboundContext =
-        obtainOutboundSsrcContext(barbell.idHash, barbell.ssrcOutboundContexts, rtxSsrc, barbell.videoFeedbackRtpMap);
-
-    if (!rtxSsrcOutboundContext)
-    {
-        return;
-    }
-
-    mediaSsrcOutboundContext->onRtpSent(timestamp);
-    const auto numFeedbackControlInfos = rtp::getNumFeedbackControlInfos(&rtcpFeedback);
-    uint16_t pid = 0;
-    uint16_t blp = 0;
-
-    auto nackPacketCount = getNackPacketCount(&rtcpFeedback, numFeedbackControlInfos);
-
-    logger::debug("Barbell received NACK for %zu pkts on ssrc %u, %s",
-        _loggableId.c_str(),
-        nackPacketCount,
-        mediaSsrc,
-        barbell.transport.getLoggableId().c_str());
-
-    for (size_t i = 0; i < numFeedbackControlInfos; ++i)
-    {
-        rtxSsrcOutboundContext->onRtpSent(timestamp);
-        rtp::getFeedbackControlInfo(&rtcpFeedback, i, numFeedbackControlInfos, pid, blp);
-        bbTransport.getJobQueue().addJob<bridge::VideoNackReceiveJob>(*rtxSsrcOutboundContext,
-            bbTransport,
-            *mediaSsrcOutboundContext,
-            pid,
-            blp,
-            timestamp,
-            barbell.transport.getRtt());
     }
 }
 
@@ -3506,27 +2615,6 @@ void EngineMixer::sendUserMediaMapMessageToAll()
     }
 }
 
-void EngineMixer::sendUserMediaMapMessageOverBarbells()
-{
-    if (_engineBarbells.size() == 0)
-    {
-        return;
-    }
-
-    utils::StringBuilder<1024> userMediaMapMessage;
-    _activeMediaList->makeBarbellUserMediaMapMessage(userMediaMapMessage, _neighbourMemberships);
-
-    if (!_engineBarbells.empty())
-    {
-        logger::debug("send BB msg %s", _loggableId.c_str(), userMediaMapMessage.get());
-    }
-
-    for (auto& barbell : _engineBarbells)
-    {
-        barbell.second->dataChannel.sendString(userMediaMapMessage.get(), userMediaMapMessage.getLength());
-    }
-}
-
 void EngineMixer::sendDominantSpeakerMessageToAll()
 {
     utils::StringBuilder<256> dominantSpeakerMessage;
@@ -3551,58 +2639,6 @@ void EngineMixer::sendDominantSpeakerMessageToAll()
         {
             sendDominantSpeakerToRecordingStream(*recStreamPair.second, dominantSpeaker, dataStream->endpointId);
         }
-    }
-}
-
-void EngineMixer::sendDominantSpeakerToRecordingStream(EngineRecordingStream& recordingStream,
-    const size_t dominantSpeaker,
-    const std::string& dominantSpeakerEndpoint)
-{
-    if (recordingStream.isVideoEnabled)
-    {
-        pinEndpoint(recordingStream.endpointIdHash, dominantSpeaker);
-    }
-
-    const auto sequenceNumber = recordingStream.recordingEventsOutboundContext.sequenceNumber++;
-    const auto timestamp = static_cast<uint32_t>(utils::Time::getAbsoluteTime() / 1000000ULL);
-
-    for (const auto& transportEntry : recordingStream.transports)
-    {
-        auto unackedPacketsTrackerItr = recordingStream.recEventUnackedPacketsTracker.find(transportEntry.first);
-        if (unackedPacketsTrackerItr == recordingStream.recEventUnackedPacketsTracker.end())
-        {
-            logger::error("RecEvent packet tracker not found. Unable to send dominant speaker recording event to %s",
-                _loggableId.c_str(),
-                transportEntry.second.getLoggableId().c_str());
-            continue;
-        }
-
-        auto packet = recp::RecDominantSpeakerEventBuilder(_sendAllocator)
-                          .setSequenceNumber(sequenceNumber)
-                          .setTimestamp(timestamp)
-                          .setDominantSpeakerEndpoint(dominantSpeakerEndpoint)
-                          .build();
-
-        if (!packet)
-        {
-            logger::warn("No space available to allocate rec dominant speaker event", _loggableId.c_str());
-            continue;
-        }
-
-        transportEntry.second.getJobQueue().addJob<RecordingSendEventJob>(std::move(packet),
-            transportEntry.second,
-            recordingStream.recordingEventsOutboundContext.packetCache,
-            unackedPacketsTrackerItr->second);
-    }
-}
-
-void EngineMixer::sendDominantSpeakerToRecordingStream(EngineRecordingStream& recordingStream)
-{
-    const auto dominantSpeaker = _activeMediaList->getDominantSpeaker();
-    auto* audioStream = _engineAudioStreams.getItem(dominantSpeaker);
-    if (audioStream)
-    {
-        sendDominantSpeakerToRecordingStream(recordingStream, dominantSpeaker, audioStream->endpointId);
     }
 }
 
@@ -3715,394 +2751,6 @@ void EngineMixer::decommissionInboundContext(const uint32_t ssrc)
     }
 }
 
-void EngineMixer::startRecordingAllCurrentStreams(EngineRecordingStream& recordingStream)
-{
-    if (recordingStream.isAudioEnabled)
-    {
-        updateRecordingAudioStreams(recordingStream, true);
-    }
-
-    if (recordingStream.isVideoEnabled)
-    {
-        updateRecordingVideoStreams(recordingStream, SimulcastStream::VideoContentType::VIDEO, true);
-        _engineStreamDirector->addParticipant(recordingStream.endpointIdHash);
-        sendDominantSpeakerToRecordingStream(recordingStream);
-    }
-
-    if (recordingStream.isScreenSharingEnabled)
-    {
-        updateRecordingVideoStreams(recordingStream, SimulcastStream::VideoContentType::SLIDES, true);
-    }
-}
-
-void EngineMixer::sendRecordingAudioStream(EngineRecordingStream& targetStream,
-    const EngineAudioStream& audioStream,
-    bool isAdded)
-{
-    const auto ssrc = audioStream.remoteSsrc.valueOr(0);
-    if (ssrc == 0)
-    {
-        return;
-    }
-
-    if (targetStream.transports.empty())
-    {
-        logger::error("Audio recording stream has zero transports. endpointHashId %zu",
-            _loggableId.c_str(),
-            targetStream.endpointIdHash);
-
-        return;
-    }
-
-    memory::UniquePacket packet;
-    auto outboundContextIt = targetStream.ssrcOutboundContexts.find(ssrc);
-    if (isAdded)
-    {
-        if (outboundContextIt != targetStream.ssrcOutboundContexts.end())
-        {
-            if (!outboundContextIt->second.recordingOutboundDecommissioned)
-            {
-                // The event already was sent
-                // It happens when audio is reconfigured
-                // We will not send the event again
-                return;
-            }
-
-            outboundContextIt->second.recordingOutboundDecommissioned = false;
-            outboundContextIt->second.markedForDeletion = false;
-
-            logger::info("Reactivating decommissioned recording outbound video ssrc %u for recording stream %zu",
-                _loggableId.c_str(),
-                ssrc,
-                targetStream.endpointIdHash);
-        }
-        else
-        {
-            auto emplaceResult =
-                targetStream.ssrcOutboundContexts.emplace(ssrc, ssrc, _sendAllocator, audioStream.rtpMap);
-
-            if (!emplaceResult.second && emplaceResult.first == targetStream.ssrcOutboundContexts.end())
-            {
-                logger::error("Failed to create outbound context for audio ssrc %u, recording stream %zu",
-                    _loggableId.c_str(),
-                    ssrc,
-                    targetStream.endpointIdHash);
-            }
-            else
-            {
-                logger::info("Created new outbound context for audio recording stream  %zu, ssrc %u",
-                    _loggableId.c_str(),
-                    targetStream.endpointIdHash,
-                    ssrc);
-            }
-        }
-
-        packet = recp::RecStreamAddedEventBuilder(_sendAllocator)
-                     .setSequenceNumber(targetStream.recordingEventsOutboundContext.sequenceNumber++)
-                     .setTimestamp(static_cast<uint32_t>(utils::Time::getAbsoluteTime() / utils::Time::ms))
-                     .setSsrc(ssrc)
-                     .setRtpPayloadType(static_cast<uint8_t>(audioStream.rtpMap.payloadType))
-                     .setPayloadFormat(audioStream.rtpMap.format)
-                     .setEndpoint(audioStream.endpointId)
-                     .setWallClock(utils::Time::now())
-                     .build();
-
-        logger::info("Audio ssrc %u added to recording %zu", _loggableId.c_str(), ssrc, targetStream.endpointIdHash);
-    }
-    else
-    {
-        if (outboundContextIt != targetStream.ssrcOutboundContexts.end())
-        {
-            outboundContextIt->second.recordingOutboundDecommissioned = true;
-        }
-
-        packet = recp::RecStreamRemovedEventBuilder(_sendAllocator)
-                     .setSequenceNumber(targetStream.recordingEventsOutboundContext.sequenceNumber++)
-                     .setTimestamp(static_cast<uint32_t>(utils::Time::getAbsoluteTime() / utils::Time::ms))
-                     .setSsrc(ssrc)
-                     .build();
-
-        logger::info("Audio ssrc %u removed from recording %zu",
-            _loggableId.c_str(),
-            ssrc,
-            targetStream.endpointIdHash);
-    }
-
-    auto lastKey = (--targetStream.transports.end())->first;
-    for (const auto& transportEntry : targetStream.transports)
-    {
-        auto unackedPacketsTrackerItr = targetStream.recEventUnackedPacketsTracker.find(transportEntry.first);
-        if (unackedPacketsTrackerItr == targetStream.recEventUnackedPacketsTracker.end())
-        {
-            logger::error("RecEvent packet tracker not found. Unable to send recording audio stream event to %s",
-                _loggableId.c_str(),
-                transportEntry.second.getLoggableId().c_str());
-            return;
-        }
-
-        memory::UniquePacket packetToSend;
-        if (lastKey == transportEntry.first)
-        {
-            packetToSend = std::move(packet);
-        }
-        else
-        {
-            packetToSend = makeUniquePacket(_sendAllocator, *packet);
-        }
-
-        if (!packetToSend)
-        {
-            // This need to be improved. If we can't allocate this event, the recording
-            // must fail as the we will not info about this stream
-            logger::error("No space to allocate rec AudioStream%s event",
-                _loggableId.c_str(),
-                isAdded ? "Added" : "Removed");
-            continue;
-        }
-
-        transportEntry.second.getJobQueue().addJob<RecordingSendEventJob>(std::move(packetToSend),
-            transportEntry.second,
-            targetStream.recordingEventsOutboundContext.packetCache,
-            unackedPacketsTrackerItr->second);
-    }
-}
-
-void EngineMixer::updateRecordingAudioStreams(EngineRecordingStream& targetStream, bool enabled)
-{
-    for (auto& audioStream : _engineAudioStreams)
-    {
-        sendRecordingAudioStream(targetStream, *audioStream.second, enabled);
-    }
-}
-
-void EngineMixer::sendRecordingVideoStream(EngineRecordingStream& targetStream,
-    const EngineVideoStream& videoStream,
-    SimulcastStream::VideoContentType contentType,
-    bool isAdded)
-{
-    if (videoStream.simulcastStream.contentType == contentType)
-    {
-        sendRecordingSimulcast(targetStream, videoStream, videoStream.simulcastStream, isAdded);
-    }
-
-    if (videoStream.secondarySimulcastStream.isSet() &&
-        videoStream.secondarySimulcastStream.get().contentType == contentType)
-    {
-        sendRecordingSimulcast(targetStream, videoStream, videoStream.secondarySimulcastStream.get(), isAdded);
-    }
-}
-
-void EngineMixer::updateRecordingVideoStreams(EngineRecordingStream& targetStream,
-    SimulcastStream::VideoContentType contentType,
-    bool enabled)
-{
-    for (auto& videoStream : _engineVideoStreams)
-    {
-        sendRecordingVideoStream(targetStream, *videoStream.second, contentType, enabled);
-    }
-}
-
-void EngineMixer::sendRecordingSimulcast(EngineRecordingStream& targetStream,
-    const EngineVideoStream& videoStream,
-    const SimulcastStream& simulcast,
-    bool isAdded)
-{
-    // The ssrc will be rewritten using level 0
-    const auto ssrc = simulcast.levels[0].ssrc;
-    if (ssrc == 0)
-    {
-        return;
-    }
-
-    if (targetStream.transports.empty())
-    {
-        logger::error("Video recording stream has zero transports. endpointHashId %zu",
-            _loggableId.c_str(),
-            targetStream.endpointIdHash);
-
-        return;
-    }
-
-    memory::UniquePacket packet;
-    if (isAdded)
-    {
-        auto outboundContextIt = targetStream.ssrcOutboundContexts.find(ssrc);
-        if (outboundContextIt != targetStream.ssrcOutboundContexts.end())
-        {
-            if (!outboundContextIt->second.recordingOutboundDecommissioned)
-            {
-                // The event already was sent
-                // It happens when audio is reconfigured
-                // We will not send the event again
-                return;
-            }
-
-            outboundContextIt->second.recordingOutboundDecommissioned = false;
-            outboundContextIt->second.markedForDeletion = false;
-
-            logger::info("Reactivating decommissioned recording outbound video ssrc %u for recording stream %zu",
-                _loggableId.c_str(),
-                ssrc,
-                targetStream.endpointIdHash);
-        }
-        else
-        {
-            auto emplaceResult =
-                targetStream.ssrcOutboundContexts.emplace(ssrc, ssrc, _sendAllocator, videoStream.rtpMap);
-
-            if (!emplaceResult.second && emplaceResult.first == targetStream.ssrcOutboundContexts.end())
-            {
-                logger::error("Failed to create outbound context for video ssrc %u, recording stream %zu",
-                    _loggableId.c_str(),
-                    ssrc,
-                    targetStream.endpointIdHash);
-            }
-            else
-            {
-                logger::info("Created new outbound context for video recording stream  %zu, ssrc %u",
-                    _loggableId.c_str(),
-                    targetStream.endpointIdHash,
-                    ssrc);
-            }
-        }
-
-        packet = recp::RecStreamAddedEventBuilder(_sendAllocator)
-                     .setSequenceNumber(targetStream.recordingEventsOutboundContext.sequenceNumber++)
-                     .setTimestamp(static_cast<uint32_t>(utils::Time::getAbsoluteTime() / utils::Time::ms))
-                     .setSsrc(ssrc)
-                     .setIsScreenSharing(simulcast.contentType == SimulcastStream::VideoContentType::SLIDES)
-                     .setRtpPayloadType(static_cast<uint8_t>(videoStream.rtpMap.payloadType))
-                     .setPayloadFormat(videoStream.rtpMap.format)
-                     .setEndpoint(videoStream.endpointId)
-                     .setWallClock(utils::Time::now())
-                     .build();
-
-        logger::info("Video ssrc %u added to recording %zu", _loggableId.c_str(), ssrc, targetStream.endpointIdHash);
-    }
-    else
-    {
-        packet = recp::RecStreamRemovedEventBuilder(_sendAllocator)
-                     .setSequenceNumber(targetStream.recordingEventsOutboundContext.sequenceNumber++)
-                     .setTimestamp(static_cast<uint32_t>(utils::Time::getAbsoluteTime() / utils::Time::ms))
-                     .setSsrc(ssrc)
-                     .build();
-
-        auto outboundContextItr = targetStream.ssrcOutboundContexts.find(ssrc);
-        if (outboundContextItr != targetStream.ssrcOutboundContexts.end())
-        {
-            outboundContextItr->second.recordingOutboundDecommissioned = true;
-        }
-
-        logger::info("Video ssrc %u removed from recording %zu",
-            _loggableId.c_str(),
-            ssrc,
-            targetStream.endpointIdHash);
-    }
-
-    auto lastKey = (--targetStream.transports.end())->first;
-    for (const auto& transportEntry : targetStream.transports)
-    {
-        auto unackedPacketsTrackerItr = targetStream.recEventUnackedPacketsTracker.find(transportEntry.first);
-        if (unackedPacketsTrackerItr == targetStream.recEventUnackedPacketsTracker.end())
-        {
-            logger::error("RecEvent packet tracker not found. Unable to send stream event to %s",
-                _loggableId.c_str(),
-                transportEntry.second.getLoggableId().c_str());
-            return;
-        }
-
-        memory::UniquePacket packetToSend;
-        if (lastKey == transportEntry.first)
-        {
-            packetToSend = std::move(packet);
-        }
-        else
-        {
-            packetToSend = makeUniquePacket(_sendAllocator, *packet);
-        }
-
-        if (!packetToSend)
-        {
-            // This need to be improved. If we can't allocate this event, the recording
-            // must fail as the we will not info about this stream
-            logger::error("No space to allocate rec VideoStream%s event",
-                _loggableId.c_str(),
-                isAdded ? "Added" : "Removed");
-            continue;
-        }
-
-        transportEntry.second.getJobQueue().addJob<RecordingSendEventJob>(std::move(packetToSend),
-            transportEntry.second,
-            targetStream.recordingEventsOutboundContext.packetCache,
-            unackedPacketsTrackerItr->second);
-    }
-}
-
-void EngineMixer::sendAudioStreamToRecording(const EngineAudioStream& audioStream, bool isAdded)
-{
-    for (auto& rec : _engineRecordingStreams)
-    {
-        if (rec.second->isAudioEnabled)
-        {
-            sendRecordingAudioStream(*rec.second, audioStream, isAdded);
-        }
-    }
-}
-
-void EngineMixer::sendVideoStreamToRecording(const EngineVideoStream& videoStream, bool isAdded)
-{
-    for (auto& rec : _engineRecordingStreams)
-    {
-        if (rec.second->isVideoEnabled)
-        {
-            sendRecordingVideoStream(*rec.second, videoStream, SimulcastStream::VideoContentType::VIDEO, isAdded);
-        }
-
-        if (rec.second->isScreenSharingEnabled)
-        {
-            sendRecordingVideoStream(*rec.second, videoStream, SimulcastStream::VideoContentType::SLIDES, isAdded);
-        }
-    }
-}
-
-void EngineMixer::removeVideoSsrcFromRecording(const EngineVideoStream& videoStream, uint32_t ssrc)
-{
-    for (auto& rec : _engineRecordingStreams)
-    {
-        if (rec.second->isVideoEnabled)
-        {
-            if (videoStream.simulcastStream.levels[0].ssrc == ssrc &&
-                videoStream.simulcastStream.contentType == SimulcastStream::VideoContentType::VIDEO)
-            {
-                sendRecordingSimulcast(*rec.second, videoStream, videoStream.simulcastStream, false);
-            }
-
-            if (videoStream.secondarySimulcastStream.isSet() &&
-                videoStream.secondarySimulcastStream.get().levels[0].ssrc == ssrc &&
-                videoStream.secondarySimulcastStream.get().contentType == SimulcastStream::VideoContentType::VIDEO)
-            {
-                sendRecordingSimulcast(*rec.second, videoStream, videoStream.simulcastStream, false);
-            }
-        }
-
-        if (rec.second->isScreenSharingEnabled)
-        {
-            if (videoStream.simulcastStream.levels[0].ssrc == ssrc &&
-                videoStream.simulcastStream.contentType == SimulcastStream::VideoContentType::SLIDES)
-            {
-                sendRecordingSimulcast(*rec.second, videoStream, videoStream.simulcastStream, false);
-            }
-
-            if (videoStream.secondarySimulcastStream.isSet() &&
-                videoStream.secondarySimulcastStream.get().levels[0].ssrc == ssrc &&
-                videoStream.secondarySimulcastStream.get().contentType == SimulcastStream::VideoContentType::SLIDES)
-            {
-                sendRecordingSimulcast(*rec.second, videoStream, videoStream.secondarySimulcastStream.get(), false);
-            }
-        }
-    }
-}
-
 void EngineMixer::processEngineMissingPackets(bridge::SsrcInboundContext& ssrcInboundContext)
 {
     auto* videoStream = _engineVideoStreams.getItem(ssrcInboundContext.sender->getEndpointIdHash());
@@ -4112,54 +2760,6 @@ void EngineMixer::processEngineMissingPackets(bridge::SsrcInboundContext& ssrcIn
             videoStream->localSsrc,
             videoStream->transport,
             _sendAllocator);
-    }
-}
-
-void EngineMixer::processBarbellMissingPackets(bridge::SsrcInboundContext& ssrcInboundContext)
-{
-    const auto barbell = _engineBarbells.getItem(ssrcInboundContext.sender->getEndpointIdHash());
-    if (barbell)
-    {
-        const uint32_t REPORTER_SSRC = 0;
-        auto videoStream = barbell->videoSsrcMap.getItem(ssrcInboundContext.ssrc);
-        if (videoStream)
-        {
-            barbell->transport.getJobQueue().addJob<bridge::ProcessMissingVideoPacketsJob>(ssrcInboundContext,
-                REPORTER_SSRC,
-                barbell->transport,
-                _sendAllocator);
-        }
-    }
-}
-
-void EngineMixer::processRecordingUnackedPackets(const uint64_t timestamp)
-{
-    if (utils::Time::diffLT(_lastRecordingAckProcessed, timestamp, RECUNACK_PROCESS_INTERVAL))
-    {
-        return;
-    }
-    _lastRecordingAckProcessed = timestamp;
-
-    for (auto& engineRecordingStreamEntry : _engineRecordingStreams)
-    {
-        auto engineRecordingStream = engineRecordingStreamEntry.second;
-        for (auto& recEventMissingPacketsTrackerEntry : engineRecordingStream->recEventUnackedPacketsTracker)
-        {
-            auto& recEventMissingPacketsTracker = recEventMissingPacketsTrackerEntry.second;
-
-            auto transportItr = engineRecordingStream->transports.find(recEventMissingPacketsTrackerEntry.first);
-            if (transportItr == engineRecordingStream->transports.end())
-            {
-                continue;
-            }
-
-            auto& transport = transportItr->second;
-            transport.getJobQueue().addJob<ProcessUnackedRecordingEventPacketsJob>(
-                engineRecordingStream->recordingEventsOutboundContext,
-                recEventMissingPacketsTracker,
-                transport,
-                _sendAllocator);
-        }
     }
 }
 
@@ -4178,21 +2778,6 @@ uint32_t EngineMixer::getMinRemoteClientDownlinkBandwidth() const
         minBitrate = std::min(minBitrate, itBb.second->minClientDownlinkBandwidth);
     }
     return minBitrate;
-}
-
-void EngineMixer::reportMinRemoteClientDownlinkBandwidthToBarbells(const uint32_t minUplinkEstimate) const
-{
-    utils::StringBuilder<1024> message;
-    api::DataChannelMessage::makeMinUplinkBitrate(message, minUplinkEstimate);
-
-    if (!_engineBarbells.empty())
-    {
-        logger::debug("send BB msg %s", _loggableId.c_str(), message.get());
-    }
-    for (const auto& itBb : _engineBarbells)
-    {
-        itBb.second->dataChannel.sendString(message.get(), message.getLength());
-    }
 }
 
 void EngineMixer::checkVideoBandwidth(const uint64_t timestamp)
@@ -4426,73 +3011,6 @@ void EngineMixer::runTransportTicks(const uint64_t timestamp)
     }
 }
 
-void EngineMixer::addBarbell(EngineBarbell* barbell)
-{
-    const auto idHash = barbell->transport.getEndpointIdHash();
-    if (_engineBarbells.contains(idHash))
-    {
-        return;
-    }
-
-    logger::debug("Add engine barbell, transport %s, idHash %lu",
-        _loggableId.c_str(),
-        barbell->transport.getLoggableId().c_str(),
-        idHash);
-
-    _engineBarbells.emplace(idHash, barbell);
-}
-
-// executed on transport thread context
-// remove inbound ssrc jobs should execute before this by calling decommissionInboundSsrc
-void EngineMixer::internalRemoveBarbell(size_t idHash)
-{
-    auto barbell = _engineBarbells.getItem(idHash);
-    if (!barbell)
-    {
-        logger::warn("barbell has already been removed from engineBarbells", _loggableId.c_str());
-        return;
-    }
-
-    _engineBarbells.erase(idHash);
-    _messageListener.asyncBarbellRemoved(*this, *barbell);
-}
-
-// executed on engine thread
-void EngineMixer::removeBarbell(size_t idHash)
-{
-    auto barbell = _engineBarbells.getItem(idHash);
-    if (!barbell)
-    {
-        return;
-    }
-
-    barbell->transport.stop();
-
-    for (const auto& videoStream : barbell->videoStreams)
-    {
-        for (auto& simulcastLevel : videoStream.stream.getLevels())
-        {
-            decommissionInboundContext(simulcastLevel.ssrc);
-            decommissionInboundContext(simulcastLevel.feedbackSsrc);
-        }
-        if (videoStream.endpointIdHash.isSet())
-        {
-            _activeMediaList->removeVideoParticipant(videoStream.endpointIdHash.get());
-        }
-    }
-
-    for (auto& audioStream : barbell->audioStreams)
-    {
-        decommissionInboundContext(audioStream.ssrc);
-        if (audioStream.endpointIdHash.isSet())
-        {
-            _activeMediaList->removeAudioParticipant(audioStream.endpointIdHash.get());
-        }
-    }
-
-    barbell->transport.postOnQueue(utils::bind(&EngineMixer::internalRemoveBarbell, this, barbell->idHash));
-}
-
 size_t EngineMixer::getDominantSpeakerId() const
 {
     return _activeMediaList->getDominantSpeaker();
@@ -4516,323 +3034,6 @@ utils::Optional<uint32_t> EngineMixer::getC9UserId(const size_t ssrc) const
 void EngineMixer::mapSsrc2UserId(uint32_t ssrc, uint32_t usid)
 {
     _audioSsrcToUserIdMap.emplace(ssrc, usid);
-}
-
-namespace
-{
-template <typename TMap>
-void copyToBarbellMapItemArray(utils::SimpleJsonArray& endpointArray, TMap& map)
-{
-    for (auto endpoint : endpointArray)
-    {
-        char endpointId[45];
-        endpoint["endpoint-id"].getString(endpointId);
-        const auto endpointIdHash = utils::hash<char*>{}(endpointId);
-        auto entryIt = map.emplace(endpointIdHash, endpointId);
-
-        auto& item = entryIt.first->second;
-        for (auto ssrc : endpoint["ssrcs"].getArray())
-        {
-            item.newSsrcs.push_back(ssrc.getInt<uint32_t>(0));
-        }
-
-        if (endpoint.exists("noise-level"))
-        {
-            item.noiseLevel = endpoint["noise-level"].getFloat(0.0);
-        }
-
-        if (endpoint.exists("level"))
-        {
-            item.noiseLevel = endpoint["level"].getFloat(0.0);
-        }
-
-        if (endpoint.exists("neighbours"))
-        {
-            for (auto neighbour : endpoint["neighbours"].getArray())
-            {
-                item.neighbours.push_back(neighbour.getInt<uint32_t>(0));
-            }
-        }
-    }
-}
-
-template <class T>
-void addToMap(EngineBarbell::VideoStream& stream, T& videoMapping)
-{
-    auto* m = videoMapping.getItem(stream.endpointIdHash.get());
-    if (!m)
-    {
-        auto entry = videoMapping.emplace(stream.endpointIdHash.get(), stream.endpointId.get().c_str());
-        if (!entry.second)
-        {
-            return;
-        }
-        m = &entry.first->second;
-    }
-
-    if (m)
-    {
-        m->oldSsrcs.push_back(stream.stream.getKeySsrc());
-    }
-}
-
-template <class T>
-void addToMap(EngineBarbell::AudioStream& stream, T& audioMapping)
-{
-    auto* m = audioMapping.getItem(stream.endpointIdHash.get());
-    if (!m)
-    {
-        auto entry = audioMapping.emplace(stream.endpointIdHash.get(), stream.endpointId.get().c_str());
-        if (!entry.second)
-        {
-            return;
-        }
-        m = &entry.first->second;
-    }
-
-    if (m)
-    {
-        m->oldSsrcs.push_back(stream.ssrc);
-    }
-}
-} // namespace
-// This method must be executed on engine thread. UMM requests could go in another queue and processed
-// at start of tick.
-// There are three phases to update the endpoints and ssrc mappings in active media list
-// Step1: Create a table of all existing stream mappings and all new stream mappings so we can compare changes
-// Step2: Remove all streams that have changed mapping and old mapping. Could be that the new mapping is nil
-// Step3: Add all streams that have changed and new mapping
-void EngineMixer::onBarbellUserMediaMap(size_t barbellIdHash, const char* message)
-{
-    auto barbell = _engineBarbells.getItem(barbellIdHash);
-    if (!barbell)
-    {
-        logger::debug("cannot find barbell for UMM. %zu", _loggableId.c_str(), barbellIdHash);
-        return;
-    }
-
-    auto mediaMapJson = utils::SimpleJson::create(message, strlen(message));
-
-    logger::info("received BB msg over barbell %s %zu, json %s",
-        _loggableId.c_str(),
-        barbell->id.c_str(),
-        barbellIdHash,
-        message);
-
-    auto videoEndpointsArray = mediaMapJson["video-endpoints"].getArray();
-    auto audioEndpointsArray = mediaMapJson["audio-endpoints"].getArray();
-
-    memory::Map<size_t, BarbellMapItem, 16> videoSsrcs;
-    copyToBarbellMapItemArray(videoEndpointsArray, videoSsrcs);
-    for (auto& stream : barbell->videoStreams)
-    {
-        if (stream.endpointIdHash.isSet())
-        {
-            addToMap(stream, videoSsrcs);
-        }
-    }
-
-    memory::Map<size_t, BarbellMapItem, 8> audioSsrcs;
-    copyToBarbellMapItemArray(audioEndpointsArray, audioSsrcs);
-    for (auto& stream : barbell->audioStreams)
-    {
-        if (stream.endpointIdHash.isSet())
-        {
-            addToMap(stream, audioSsrcs);
-        }
-    }
-
-    const auto mapRevision = _activeMediaList->getMapRevision();
-
-    // remove video
-    for (const auto& entry : videoSsrcs)
-    {
-        const auto& item = entry.second;
-        if (item.hasChanged() && !item.oldSsrcs.empty())
-        {
-            _activeMediaList->removeVideoParticipant(entry.first);
-            _engineStreamDirector->removeParticipant(entry.first);
-
-            for (const auto ssrc : item.oldSsrcs)
-            {
-                auto* videoStream = barbell->videoSsrcMap.getItem(ssrc);
-                if (videoStream)
-                {
-                    videoStream->endpointIdHash.clear();
-                    videoStream->endpointId.clear();
-                    videoStream->stream.highestActiveLevel = 0;
-                }
-            }
-        }
-    }
-
-    // add videos
-    for (const auto& entry : videoSsrcs)
-    {
-        const auto& item = entry.second;
-        if (item.hasChanged() && !item.newSsrcs.empty())
-        {
-            SimulcastStream primary;
-            utils::Optional<SimulcastStream> secondary;
-            uint32_t streamIndex = 0;
-            for (const auto ssrc : item.newSsrcs)
-            {
-                auto* videoStream = barbell->videoSsrcMap.getItem(ssrc);
-                if (!videoStream)
-                {
-                    assert(false);
-                    logger::error("unannounced video ssrc %u", _loggableId.c_str(), ssrc);
-                    continue;
-                }
-
-                videoStream->stream.highestActiveLevel = 0;
-
-                if (streamIndex++ == 0)
-                {
-                    primary = videoStream->stream;
-                }
-                else
-                {
-                    secondary.set(videoStream->stream);
-                }
-                videoStream->endpointIdHash.set(entry.first);
-                videoStream->endpointId.set(item.endpointId);
-            }
-
-            _activeMediaList->addBarbellVideoParticipant(entry.first, primary, secondary, item.endpointId);
-            _engineStreamDirector->addParticipant(entry.first, primary, secondary.isSet() ? &secondary.get() : nullptr);
-
-            for (const auto& simLevel : primary.getLevels())
-            {
-                auto* inboundContext = _ssrcInboundContexts.getItem(simLevel.ssrc);
-                if (inboundContext)
-                {
-                    inboundContext->makeReady();
-                    inboundContext->pliScheduler.triggerPli();
-                }
-            }
-            if (secondary.isSet())
-            {
-                for (const auto& simLevel : secondary.get().getLevels())
-                {
-                    auto* inboundContext = _ssrcInboundContexts.getItem(simLevel.ssrc);
-                    if (inboundContext)
-                    {
-                        inboundContext->makeReady();
-                        inboundContext->pliScheduler.triggerPli();
-                    }
-                }
-            }
-        }
-    }
-
-    const auto audioMapRevision = _activeMediaList->getMapRevision();
-
-    // remove audio
-    for (const auto& entry : audioSsrcs)
-    {
-        const auto& item = entry.second;
-        if (item.hasChanged() && !item.oldSsrcs.empty())
-        {
-            _activeMediaList->removeAudioParticipant(entry.first);
-            _neighbourMemberships.erase(entry.first);
-            for (const auto ssrc : item.oldSsrcs)
-            {
-                auto* audioStream = barbell->audioSsrcMap.getItem(ssrc);
-                if (audioStream)
-                {
-                    audioStream->endpointIdHash.clear();
-                    audioStream->endpointId.clear();
-                }
-            }
-        }
-    }
-
-    // add audio
-    for (const auto& entry : audioSsrcs)
-    {
-        const auto& item = entry.second;
-        if (item.hasChanged() && !item.newSsrcs.empty())
-        {
-            auto* audioStream = barbell->audioSsrcMap.getItem(item.newSsrcs[0]);
-            if (!audioStream)
-            {
-                logger::error("unannounced audio ssrc %u", _loggableId.c_str(), item.newSsrcs[0]);
-                // if it is set it must already be set to this endpointId
-                continue;
-            }
-
-            audioStream->endpointIdHash.set(entry.first);
-            audioStream->endpointId.set(item.endpointId);
-            _activeMediaList->addBarbellAudioParticipant(entry.first,
-                item.endpointId,
-                item.noiseLevel,
-                item.recentLevel);
-            if (!item.neighbours.empty())
-            {
-                auto neighbourIt = _neighbourMemberships.emplace(entry.first, entry.first);
-                if (neighbourIt.second)
-                {
-                    auto& newMembershipItem = neighbourIt.first->second;
-                    utils::append(newMembershipItem.memberships, item.neighbours);
-                }
-            }
-        }
-    }
-
-    if (logger::_logLevel >= logger::Level::DBG && audioMapRevision != _activeMediaList->getMapRevision())
-    {
-        _activeMediaList->logAudioList();
-    }
-
-    if (mapRevision != _activeMediaList->getMapRevision())
-    {
-        // only update local users
-        sendUserMediaMapMessageToAll();
-    }
-}
-
-void EngineMixer::onBarbellMinUplinkEstimate(size_t barbellIdHash, const char* message)
-{
-    auto barbell = _engineBarbells.getItem(barbellIdHash);
-    if (!barbell)
-    {
-        return;
-    }
-
-    logger::debug("received BB msg over barbell %s %zu, json %s",
-        _loggableId.c_str(),
-        barbell->id.c_str(),
-        barbellIdHash,
-        message);
-
-    barbell->minClientDownlinkBandwidth =
-        api::DataChannelMessageParser::getMinUplinkBitrate(utils::SimpleJson::create(message, strlen(message)));
-}
-
-void EngineMixer::onBarbellDataChannelEstablish(size_t barbellIdHash,
-    webrtc::SctpStreamMessageHeader& header,
-    size_t packetSize)
-{
-    auto barbell = _engineBarbells.getItem(barbellIdHash);
-    if (!barbell)
-    {
-        return;
-    }
-
-    const auto state = barbell->dataChannel.getState();
-    barbell->dataChannel.onSctpMessage(&barbell->transport,
-        header.id,
-        header.sequenceNumber,
-        header.payloadProtocol,
-        header.data(),
-        header.getMessageLength(packetSize));
-
-    const auto newState = barbell->dataChannel.getState();
-    if (state != newState && newState == webrtc::WebRtcDataStream::State::OPEN)
-    {
-        sendUserMediaMapMessageOverBarbells();
-    }
 }
 
 bool EngineMixer::asyncReconfigureVideoStream(const transport::RtcTransport& transport,
@@ -4916,86 +3117,9 @@ bool EngineMixer::asyncSendEndpointMessage(const size_t toEndpointIdHash,
         utils::moveParam(packet)));
 }
 
-bool EngineMixer::asyncAddRecordingStream(EngineRecordingStream* engineRecordingStream)
-{
-    return post(utils::bind(&EngineMixer::addRecordingStream, this, engineRecordingStream));
-}
-
-bool EngineMixer::asyncAddTransportToRecordingStream(const size_t streamIdHash,
-    transport::RecordingTransport& transport,
-    UnackedPacketsTracker& recUnackedPacketsTracker)
-{
-    return post(utils::bind(&EngineMixer::addTransportToRecordingStream,
-        this,
-        streamIdHash,
-        std::ref(transport),
-        std::ref(recUnackedPacketsTracker)));
-}
-
-bool EngineMixer::asyncRecordingStart(EngineRecordingStream& stream, const RecordingDescription& desc)
-{
-    return post(utils::bind(&EngineMixer::recordingStart, this, std::ref(stream), desc));
-}
-
-bool EngineMixer::asyncUpdateRecordingStreamModalities(EngineRecordingStream& engineRecordingStream,
-    bool isAudioEnabled,
-    bool isVideoEnabled,
-    bool isScreenSharingEnabled)
-{
-    return post(utils::bind(&EngineMixer::updateRecordingStreamModalities,
-        this,
-        std::ref(engineRecordingStream),
-        isAudioEnabled,
-        isVideoEnabled,
-        isScreenSharingEnabled));
-}
-
-bool EngineMixer::asyncStartRecordingTransport(transport::RecordingTransport& transport)
-{
-    return post(utils::bind(&EngineMixer::startRecordingTransport, this, std::ref(transport)));
-}
-
-bool EngineMixer::asyncStopRecording(EngineRecordingStream& stream, const RecordingDescription& desc)
-{
-    return post([this, &stream, &desc]() {
-        this->stopRecording(stream, desc);
-        _messageListener.asyncEngineRecordingStopped(*this, desc);
-    });
-}
-
-bool EngineMixer::asyncAddRecordingRtpPacketCache(const uint32_t ssrc,
-    const size_t endpointIdHash,
-    PacketCache* packetCache)
-{
-    return post(utils::bind(&EngineMixer::addRecordingRtpPacketCache, this, ssrc, endpointIdHash, packetCache));
-}
-
-bool EngineMixer::asyncRemoveTransportFromRecordingStream(const size_t streamIdHash, const size_t endpointIdHash)
-{
-    return post(utils::bind(&EngineMixer::removeTransportFromRecordingStream, this, streamIdHash, endpointIdHash));
-}
-
-bool EngineMixer::asyncAddBarbell(EngineBarbell* barbell)
-{
-    return post(utils::bind(&EngineMixer::addBarbell, this, barbell));
-}
-
-bool EngineMixer::asyncRemoveBarbell(size_t idHash)
-{
-    return post(utils::bind(&EngineMixer::removeBarbell, this, idHash));
-}
-
 bool EngineMixer::asyncHandleSctpControl(const size_t endpointIdHash, memory::UniquePacket& packet)
 {
     return post(utils::bind(&EngineMixer::handleSctpControl, this, endpointIdHash, utils::moveParam(packet)));
-}
-
-bool EngineMixer::asyncRemoveRecordingStream(const EngineRecordingStream& engineRecordingStream)
-{
-    return post([this, &engineRecordingStream]() {
-        this->removeRecordingStream(engineRecordingStream);
-        _messageListener.asyncRecordingStreamRemoved(*this, engineRecordingStream);
-    });
 }
 
 } // namespace bridge
